refactor: modularize batch processing
