+++ package.json
@@ -9,10 +9,7 @@
     "start": "npm run build && node dist/index.js"
   },
   "dependencies": {
-    "body-parser": "^1.20.2",
     "chartjs-node-canvas": "^5.0.0",
-    "cors": "^2.8.5",
-    "express": "^4.18.2",
     "ws": "^8.18.3"
   },
   "repository": {
